prescription_image_ocr/requirements.txt
fastapi
uvicorn
pytesseract
pillow
opencv-python-headless
rapidfuzz
neo4j
python-multipart
